+++ R/loo_subsample.R
@@ -28,6 +28,10 @@
 #' @export
 loo_subsample.brmsfit <- function(x, ..., compare = TRUE, resp = NULL,
                                   model_names = NULL) {
+  if (has_mix_groups(x$family)) {
+    stop2("'loo_subsample' is not supported for group-level mixture ",
+          "models (specified via 'gr' in mixture()).")
+  }
   args <- split_dots(x, ..., model_names = model_names)
   c(args) <- nlist(
     criterion = "loo_subsample", compare, resp,